requirements/requirements.txt
apache-libcloud == 2.4.0 --hash=sha256:125c410996b84464b426922f1398a317869f27173a6461e32f3b1dfe671d5235
certifi == 2018.11.29 --hash=sha512:6a6bf1ff98caefcdbf78a8c83e11e155368bacdd806f0ae0c6afa8f513667df6598e594b3584de61acdca3d6049f4a776937f2aa8672b602bd6db7b737f6074e
chardet == 3.0.4 --hash=sha256:84ab92ed1c4d4f16916e05906b6b75a6c0fb5db821cc65e70cbd64a3e2a5eaae
click == 6.7 --hash=sha512:d454369fe1acb01c2210a61e15f45ac9cc77102c93345257f4ef3090bd8f31f1ccc49d9fd6ea670cddca9f715ad5db29368b4eecd95abff50d1033c4f7a967d6
idna == 2.8 --hash=sha512:8ca5cfe6350c51250bafdac7c6e4ddd54c4a5d6bf7acbcef896760a759868c8e9df1fdf550121d8512fa3eb316dcf031ec6058e03b4f66eadee21b63e2187d33
munch == 2.2.0 --hash=sha256:62fb4fb318e965a464b088e6af52a63e0905a50500b770596a939d3855e7aa15
requests == 2.21.0 --hash=sha256:502a824f31acdacb3a35b6690b5fbf0bc41d63a24a45c4004352b0242707598e
six == 1.12.0 --hash=sha512:937728372edf1e0ac13bbd706723d0de35e015c30d0ae41f789c5ed2e3669bb0db70cdc6e036ec2d437a6c4aa0d5d1e727b6d09ac34cca7e4e92e5d3b4775151
texttable == 1.6.0 --hash=sha256:624b9d6a94255b3d168b22f0808b78eca8b765d8b14c095e530641b8655c7dfc
toml == 0.10.0 --hash=sha512:26f26c38ce9cd48305218c2c34c5a5407b00aefc25a933f044bb7be22c23cfdfa3b8cf2da952d17760c4b9038df62e405fa039cc7b63dd3e94c9c312f04f9182
urllib3 == 1.24.1 --hash=sha256:de9529817c93f27c8ccbfead6985011db27bd0ddfcdb2d86f3f663385c6a9c22
